requirements.txt
appdirs
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
packaging
protobuf
pyparsing
six
torch
scipy==1.1.0
Pillow
keras
h5py
